Reverted commit: "Changed all weights from doubles to integers"

+++ app/src/main/java/com/example/fitnessapp/DefaultWeights.java
@@ -7,25 +7,25 @@ import java.util.List;
 public class DefaultWeights {
 
     //Default weights that are assigned when checkbox for exercises in NewProgramActivity are unchecked
-    public static final int BENCH_PRESS_BEGINNER = 65;
-    public static final int OVERHEAD_PRESS_BEGINNER = 45;
-    public static final int SQUAT_BEGINNER = 95;
-    public static final int DEADLIFT_BEGINNER = 135;
-    public static final int BARBELL_ROW_BEGINNER = 45;
+    public static final double BENCH_PRESS_BEGINNER = 65;
+    public static final double OVERHEAD_PRESS_BEGINNER = 45;
+    public static final double SQUAT_BEGINNER = 95;
+    public static final double DEADLIFT_BEGINNER = 135;
+    public static final double BARBELL_ROW_BEGINNER = 45;
 
-    public static final int BENCH_PRESS_INTERMEDIATE = 165;
-    public static final int OVERHEAD_PRESS_INTERMEDIATE = 105;
-    public static final int SQUAT_INTERMEDIATE = 225;
-    public static final int DEADLIFT_INTERMEDIATE = 250;
-    public static final int BARBELL_ROW_INTERMEDIATE = 135;
+    public static final double BENCH_PRESS_INTERMEDIATE = 165;
+    public static final double OVERHEAD_PRESS_INTERMEDIATE = 105;
+    public static final double SQUAT_INTERMEDIATE = 225;
+    public static final double DEADLIFT_INTERMEDIATE = 250;
+    public static final double BARBELL_ROW_INTERMEDIATE = 135;
 
-    public static final int BENCH_PRESS_ADVANCED = 225;
-    public static final int OVERHEAD_PRESS_ADVANCED = 135;
-    public static final int SQUAT_ADVANCED = 315;
-    public static final int DEADLIFT_ADVANCED = 350;
-    public static final int BARBELL_ROW_ADVANCED = 185;
+    public static final double BENCH_PRESS_ADVANCED = 225;
+    public static final double OVERHEAD_PRESS_ADVANCED = 135;
+    public static final double SQUAT_ADVANCED = 315;
+    public static final double DEADLIFT_ADVANCED = 350;
+    public static final double BARBELL_ROW_ADVANCED = 185;
 
-    public static final List<Integer> defaultWeights = new ArrayList<>(Arrays.asList(
+    public static final List<Double> defaultWeights = new ArrayList<>(Arrays.asList(
             BENCH_PRESS_BEGINNER,
             OVERHEAD_PRESS_BEGINNER,
             SQUAT_BEGINNER,
