+++ app/src/main/java/com/example/fitnessapp/activities/BaseWorkoutLogActivity.java
@@ -35,7 +35,7 @@ public abstract class BaseWorkoutLogActivity extends AppCompatActivity {
         BARBELL ROW
      */
 
-    private int[] exerciseWeights;
+    private double[] exerciseWeights;
     private String[] exerciseNames;
 
     private int increment;
@@ -151,11 +151,11 @@ public abstract class BaseWorkoutLogActivity extends AppCompatActivity {
             //Checking to see if all EditTexts are filled
             if (areAllFilled(weightsInput, repsInput)) {
                 try {
-                    int[] weights = new int[currentWorkout.getExercises().size()];
+                    double[] weights = new double[currentWorkout.getExercises().size()];
                     int[] reps = new int[currentWorkout.getExercises().size()];
 
                     for (int i = 0; i < weights.length; i++) {
-                        weights[i] = Integer.parseInt(weightsInput[i]);
+                        weights[i] = Double.parseDouble(weightsInput[i]);
                         reps[i] = Integer.parseInt(repsInput[i]);
                     }
 
@@ -195,11 +195,11 @@ public abstract class BaseWorkoutLogActivity extends AppCompatActivity {
                     }
 
                     //Gets new goal weight
-                    int capableWeight = exercise.getCapableWeight();
+                    double capableWeight = exercise.getCapableWeight();
 
                     //Checks if database contains any entries with the current time of today and workoutExerciseID
                     if (databaseHelper.haveEntriesBeenEntered(todaysTime, workoutExerciseID)) {
-                        List<Integer> weightsForDataTable = new ArrayList<>();
+                        List<Double> weightsForDataTable = new ArrayList<>();
                         List<Integer> repsForDataTable = new ArrayList<>();
 
                         for (int i = 0; i < currentWorkout.getExercises().size(); i++) {
@@ -237,7 +237,7 @@ public abstract class BaseWorkoutLogActivity extends AppCompatActivity {
         percentage = perc;
     }
 
-    protected void setExerciseWeights(int[] weights) {
+    protected void setExerciseWeights(double[] weights) {
         exerciseWeights = weights;
     }
 
@@ -278,7 +278,7 @@ public abstract class BaseWorkoutLogActivity extends AppCompatActivity {
     }
 
     //Rounds exercise weight to nearest multiple of 5
-    protected int round(double weight) {
+    protected double round(double weight) {
         return 5 * Math.round((int) (weight / 5));
     }
 
@@ -319,7 +319,7 @@ public abstract class BaseWorkoutLogActivity extends AppCompatActivity {
         return !weight.isShown() && !reps.isShown();
     }
 
-    private void addRepsDoneToExercise(Exercise exercise, int[] weights, int[] reps) {
+    private void addRepsDoneToExercise(Exercise exercise, double[] weights, int[] reps) {
         for (int i = 0; i < weights.length; i++) {
             exercise.addRepsDone(weights[i], reps[i]);
         }
